address review comments

+++ multi-environments-kustomize/cleanup.sh
@@ -33,6 +33,11 @@ if [[ -z "$PROD_CLUSTER_ZONE" ]]; then
     exit 1
 fi
 
+if [[ -z "$CM_CONFIG_DIR" ]]; then
+    echo "Must provide CM_CONFIG_DIR in environment" 1>&2
+    exit 1
+fi
+
 echo "Turning off Anthos Config Management.."
 gcloud config set project $DEV_PROJECT
 gcloud alpha container hub config-management disable 
@@ -45,36 +50,36 @@ echo "Deleting GKE clusters..."
 gcloud container clusters delete --quiet --project=${DEV_PROJECT} dev --zone=${DEV_CLUSTER_ZONE} --async 
 gcloud container clusters delete --quiet --project=${PROD_PROJECT} prod --zone=${PROD_CLUSTER_ZONE} --async
 
-echo "Deleting Secret Manager secrets..."
-gcloud config set project $PROD_PROJECT
-gcloud secrets delete github-username --quiet
-gcloud secrets delete github-email --quiet 
-gcloud secrets delete github-token --quiet
+if [[ $CM_CONFIG_DIR == "cloud-build-rendering" ]]; then
+    echo "Deleting Secret Manager secrets..."
+    gcloud config set project $PROD_PROJECT
+    gcloud secrets delete github-username --quiet
+    gcloud secrets delete github-email --quiet
+    gcloud secrets delete github-token --quiet
 
+    echo "Deleting foo repos from GitHub..."
+    curl \
+      -X DELETE \
+      -H "Accept: application/vnd.github.v3+json" \
+      -H "Authorization: token ${GITHUB_TOKEN}" \
+       "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-source"
 
-echo "Deleting foo repos from GitHub..."
-curl \
-  -X DELETE \
-  -H "Accept: application/vnd.github.v3+json" \
-  -H "Authorization: token ${GITHUB_TOKEN}" \
-   "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-source"
+    curl \
+      -X DELETE \
+      -H "Accept: application/vnd.github.v3+json" \
+      -H "Authorization: token ${GITHUB_TOKEN}" \
+       "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-prod"
 
-curl \
-  -X DELETE \
-  -H "Accept: application/vnd.github.v3+json" \
-  -H "Authorization: token ${GITHUB_TOKEN}" \
-   "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-prod"
+    curl \
+      -X DELETE \
+      -H "Accept: application/vnd.github.v3+json" \
+      -H "Authorization: token ${GITHUB_TOKEN}" \
+       "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-dev"
 
-curl \
-  -X DELETE \
-  -H "Accept: application/vnd.github.v3+json" \
-  -H "Authorization: token ${GITHUB_TOKEN}" \
-   "https://api.github.com/repos/$GITHUB_USERNAME/foo-config-dev"
-
-
-echo "Removing local files..."
-rm -r dev-key.json 
-rm -r prod-key.json 
-rm -rf foo-config-source/
-rm -rf foo-config-dev/
-rm -rf foo-config-prod/
+    echo "Removing local files..."
+    rm -r dev-key.json
+    rm -r prod-key.json
+    rm -rf foo-config-source/
+    rm -rf foo-config-dev/
+    rm -rf foo-config-prod/
+fi
